+++ packages/stylelint-config-itgalaxy/lib/config/rules/media-feature.js
@@ -7,6 +7,7 @@ module.exports = {
     'media-feature-colon-space-before': 'never',
     'media-feature-name-no-vendor-prefix': true,
     'media-feature-no-missing-punctuation': true,
+    'media-feature-parentheses-space-inside': 'never',
     'media-feature-range-operator-space-after': 'always',
     'media-feature-range-operator-space-before': 'always'
 };