+++ packages/stylelint-config-itgalaxy/lib/config/rules/property.js
@@ -5,6 +5,7 @@
 module.exports = {
     'property-blacklist': null,
     'property-case': 'lower',
+    'property-no-unknown': true,
     'property-no-vendor-prefix': true,
     'property-whitelist': null
 };